Get basic bellows functionality working again

+++ modules/patches/Playlist.js
@@ -2,19 +2,19 @@ import {overrideFunc} from './patcher.js'
 import { getApi } from '../apis/index.js';
 
 Playlist.prototype.findOrCreatePlayer = function(sound) {
-	if (sound.flags.bIsStreamed && sound.flags.streamingApi !== undefined)
+	if (hasProperty(sound, "data.flags.bIsStreamed") && hasProperty(sound, "data.flags.streamingApi") && sound.data.flags.bIsStreamed && sound.data.flags.streamingApi !== undefined)
 	{
-		return getApi(sound.flags.streamingApi).findOrCreatePlayer(
-			this.id, sound._id, sound.flags.streamingId
+		return getApi(sound.data.flags.streamingApi).findOrCreatePlayer(
+			this.id, sound.data._id, sound.data.flags.streamingId
 		);
 	}
 	return null;
 };
 
 Playlist.prototype.cleanupPlayer = function (sound) {
-	if (sound.flags.bIsStreamed && sound.flags.streamingApi !== undefined)
+	if (hasProperty(sound, "data.flags.bIsStreamed") && hasProperty(sound, "data.flags.streamingApi") && sound.data.flags.bIsStreamed && sound.data.flags.streamingApi !== undefined)
 	{
-		getApi(sound.flags.streamingApi).cleanupPlayer(
+		getApi(sound.data.flags.streamingApi).cleanupPlayer(
 			this.id, sound._id
 		);
 	}
@@ -28,38 +28,62 @@ Playlist.prototype.cleanupPlayer = function (sound) {
  */
 overrideFunc(Playlist.prototype, '_createAudio', function(super_createAudio, sound)
 {
-	if (!sound.flags.bIsStreamed)
+	if (!hasProperty(sound, "data.flags.bIsStreamed") || !sound.data.flags.bIsStreamed)
 	{
 		super_createAudio.call(this, sound);
-	} 
+	}
 	else if(sound.playing)
 	{
 		//resume after foundry suspension
 		let player = this.findOrCreatePlayer(sound);
-		player.setSourceId(sound.flags.streamingId);
+		player.setSourceId(sound.data.flags.streamingId);
 		player.setLoop(sound.repeat);
 		player.setVolume(sound.volume * game.settings.get("core", "globalPlaylistVolume"));
 		player.ensurePlaying(sound.playing);
 	}
 });
 
-overrideFunc(Playlist.prototype, 'playSound', function(super_playSound, sound)
+overrideFunc(Playlist.prototype, 'stopSound', function(super_stopSound, sound)
 {
-	if (!sound.flags.bIsStreamed)
+	if (!hasProperty(sound.data, "flags.bIsStreamed") || !sound.data.flags.bIsStreamed)
 	{
-		super_playSound.call(this, sound);
+		super_stopSound.call(this, sound);
 		return;
 	}
 
 	if (sound.playing) {
-
 		let player = this.findOrCreatePlayer(sound);
-		player.setSourceId(sound.flags.streamingId);
+		player.stopPlaying()
+        super_stopSound.call(this, sound);
+		console.log('playSound', sound.data.flags.streamingId, sound.playing);
+
+	} else {
+		this.cleanupPlayer(sound);
+	}
+});
+
+overrideFunc(Playlist.prototype, 'playSound', function(super_playSound, sound)
+{
+	if (!hasProperty(sound.data, "flags.bIsStreamed") || !sound.data.flags.bIsStreamed)
+	{
+	    console.log("playSound value of bIsStreamed=" + hasProperty(sound.data, "sound.data.flags.bIsStreamed"))
+		super_playSound.call(this, sound);
+		console.log("calling super playerSound")
+		return;
+	}
+
+
+    let player = this.findOrCreatePlayer(sound);
+	if (player && !player.isPlaying()) {
+	    const updates = {playing: true};
+		player.setSourceId(sound.data.flags.streamingId);
 		player.setLoop(sound.repeat);
 		player.setVolume(sound.volume * game.settings.get("core", "globalPlaylistVolume"));
-		player.ensurePlaying(sound.playing);
-		console.log('playSound', sound.flags.streamingId, sound.playing);
-		
+		updates.sounds = [{_id: sound.id, playing: true}];
+		this.update(updates);
+		player.ensurePlaying(true);
+		console.log('playSound', sound.data.flags.streamingId, sound.playing);
+
 	} else {
 		this.cleanupPlayer(sound);
 	}
@@ -72,7 +96,7 @@ overrideFunc(Playlist.prototype, '_onDeleteEmbeddedEntity', function(
 {
 	super_onDeleteEmbeddedEntity.call(this, embeddedName, child, options, userId);
 	console.log('Deleting', child);
-	if (child.flags.bIsStreamed)
+	if (hasProperty(child, "data.flags.bIsStreamed") && child.data.flags.bIsStreamed)
 	{
 		this.findOrCreatePlayer(child).delete();
 	}
